requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.0
asn1crypto==1.5.1
base58==2.1.1
bip-utils==2.9.3
bitarray==2.9.2
blinker==1.8.2
boto3==1.35.23
botocore==1.35.23
cbor2==5.6.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
coincurve==20.0.0
construct==2.10.68
construct-typing==0.5.6
crcmod==1.7
cytoolz==0.12.3
ecdsa==0.19.0
ed25519==1.5
ed25519-blake2b==1.4.1
eth-account==0.8.0
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
eth_abi==5.1.0
Flask==3.0.3
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonalias==0.1.1
MarkupSafe==2.1.5
mnemonic==0.21
mpmath==1.3.0
msgpack==1.1.0
mysql-connector-python==9.0.0
packaging==24.1
parsimonious==0.10.0
peewee==3.17.6
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
solana==0.34.3
solders==0.21.0
sympy==1.13.2
toolz==0.12.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
websockets==11.0.3
Werkzeug==3.0.4
